feature: modified an example.
