+++ crates/core/src/agent/executor.rs
@@ -437,4 +437,5 @@ mod tests {
         let error = TestError::TestError("source test".to_string());
         assert!(error.source().is_none());
     }
+    
 }
