+++ crates/llm/src/lib.rs
@@ -83,6 +83,7 @@ pub struct FunctionCall {
     pub arguments: String,
 }
 
+
 #[cfg(test)]
 mod tests {
     use super::*;
