+++ examples/streaming_real/Cargo.toml
